refactor: 프로젝트 관리 호스트 컴포넌트 정리
