+++ src/components/project/management/host/recruiting/applicant-list.tsx
@@ -4,8 +4,8 @@ import useDisclosure from '@/hooks/useDisclosure';
 import { IApplicant } from '@/lib/api/project/project.interface';
 import { cn } from '@/lib/utils';
 import { IProject } from '@/types/project.type';
-import Icon from '../../common/icon';
-import { RecruitingProjectApplyGuestItem } from '../apply-member-item';
+import Icon from '../../../../common/icon';
+import { RecruitingProjectApplyGuestItem } from '../../../apply-member-item';
 
 export const ApplicantList = ({
   projectId,
